backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
pillow
torch
torchvision
requests